requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
bcrypt==5.0.0
blis==1.3.3
catalogue==2.0.10
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
cloudpathlib==0.23.0
confection==1.3.3
cryptography==46.0.7
cymem==2.0.13
dateparser==1.4.0
distro==1.9.0
dnspython==2.8.0
ecdsa==0.19.2
email-validator==2.3.0
fastapi==0.117.1
google-api-core==2.30.3
google-api-python-client==2.194.0
google-auth==2.49.2
google-auth-httplib2==0.3.1
google-auth-oauthlib==1.3.1
googleapis-common-protos==1.74.0
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.2
httpx==0.28.1
idna==3.11
iniconfig==2.3.0
jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
markdown-it-py==4.0.0
markupsafe==3.0.3
mdurl==0.1.2
murmurhash==1.0.15
nltk==3.9.4
numpy==2.4.4
oauthlib==3.3.1
openai==1.109.1
packaging==26.0
passlib[bcrypt]==1.7.4
pendulum==3.2.0
pluggy==1.6.0
preshed==3.0.13
proto-plus==1.27.2
protobuf==7.34.1
psycopg2-binary==2.9.11
pyasn1==0.6.3
pyasn1-modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.13.1
pygments==2.20.0
pyparsing==3.3.2
pytest==8.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-jose[cryptography]==3.5.0
python-multipart
pytz==2026.1.post1
regex==2026.4.4
requests==2.33.1
requests-oauthlib==2.0.0
rich==14.3.3
rsa==4.9.1
shellingham==1.5.4
six==1.17.0
smart-open==7.5.1
sniffio==1.3.1
spacy==3.8.14
spacy-legacy==3.0.12
spacy-loggers==1.0.5
sqlalchemy==2.0.49
srsly==2.5.3
starlette==0.48.0
thinc==8.3.13
tqdm==4.67.3
typer==0.24.1
typing-extensions==4.15.0
typing-inspection==0.4.2
tzdata==2026.1
tzlocal==5.3.1
uritemplate==4.2.0
urllib3==2.6.3
uvicorn==0.37.0
wasabi==1.1.3
weasel==1.0.0
wrapt==2.1.2
caldav>=1.3.0
